requirements.txt
pyspark
injector
python-dotenv
pandas
numpy
matplotlib
apache-superset